+++ scout/adapter/mongo/case_events.py
@@ -81,9 +81,17 @@ class CaseEventHandler(object):
         LOG.info("Updating {0} to be unassigned with {1}".format(
             case['display_name'], user['name']))
 
+        # if no other user is assigned to the case and case is not prioritized
+        if case['status'] != 'prioritized' and case.get('assignees') == [user['email']]:
+            # flag case as inactive:
+            case['status'] = 'inactive'
+
         updated_case = self.case_collection.find_one_and_update(
             {'_id': case['_id']},
-            {'$pull': {'assignees': user['_id']}},
+            {
+                '$pull': {'assignees': user['_id']},
+                '$set': { 'status': case['status']}
+            },
             return_document=pymongo.ReturnDocument.AFTER
         )
         LOG.debug("Case updated")
@@ -115,6 +123,11 @@ class CaseEventHandler(object):
         LOG.info("Creating event for updating status of {0} to {1}".format(
             case['display_name'], status))
 
+        # assign case to user if user unarchives it
+        if case.get('status') == 'archived' and status == 'active':
+            LOG.info('assign case to user {}'.format(user['email']))
+            self.assign(institute, case, user, link)
+
         self.create_event(
             institute=institute,
             case=case,
@@ -132,6 +145,7 @@ class CaseEventHandler(object):
             return_document=pymongo.ReturnDocument.AFTER
         )
         LOG.debug("Case updated")
+
         return updated_case
 
     def update_synopsis(self, institute, case, user, link, content=""):
@@ -261,6 +275,10 @@ class CaseEventHandler(object):
         if case.get('rerun_requested'):
             raise ValueError('rerun already pending')
 
+        if case.get('status') == 'archived':
+            # assign case to user requesting rerun
+            self.assign(institute, case, user, link)
+
         self.create_event(
             institute=institute,
             case=case,
@@ -274,7 +292,10 @@ class CaseEventHandler(object):
         updated_case = self.case_collection.find_one_and_update(
             {'_id': case['_id']},
             {
-                '$set': {'rerun_requested': True}
+                '$set': {
+                    'rerun_requested': True,
+                    'status': case.get('status') if case.get('status') != 'archived' else 'inactive'
+                    }
             },
             return_document=pymongo.ReturnDocument.AFTER
         )
@@ -478,7 +499,7 @@ class CaseEventHandler(object):
         )
         LOG.debug("Case updated")
         return updated_case
-    
+
     def mark_checked(self, institute, case, user, link,
                      unmark=False):
         """Mark a case as checked from an analysis point of view.
@@ -520,7 +541,7 @@ class CaseEventHandler(object):
         )
         LOG.debug("Case updated")
         return updated_case
-    
+
 
     def update_default_panels(self, institute_obj, case_obj, user_obj, link, panel_objs):
         """Update default panels for a case.